requirements.txt
dotenv
markdown-it-py